Refactor binary tree.
